requirements.txt
farm-haystack[ocr]==1.15.1
farm-haystack[preprocessing]==1.15.1
rouge-score==0.1.2 
sentence-transformers==2.2.2 
sentencepiece==0.1.98 
tokenizers==0.13.3 
evaluate==0.4.0
datasets==2.11.0
nltk==3.8.1
torch == 1.12.1
sacrebleu == 2.2.1
jsonlines == 3.1.0
absl-py == 1.2.0
accelerate == 0.18.0